refactor: remove scheduler_py and resolve imports
