--- scheduler_py/setup.cfg
@@ -1,67 +1,0 @@
-[tool:pytest]
-testpaths = "tests"
-pythonpath = src
-addopts =
-  --strict-markers
-  --tb=short
-  --cov=src
-  --cov-branch
-  --cov-report=term-missing
-  --cov-report=html
-  --cov-report=xml
-  --no-cov-on-fail
-  --cov-fail-under=90
-
-
-[coverage:report]
-precision = 2
-exclude_lines =
-    pragma: no cover
-    raise NotImplementedError
-    raise NotImplemented
-
-[coverage:run]
-source = src
-branch = True
-omit =
- src/main.py
-
-[isort]
-profile = "black"
-src_paths =
-  "src",
-  "tests"
-combine_as_imports = true
-
-[flake8]
-max-line-length = 120
-exclude =
-    migrations
-ignore =
-    # common errors:
-    # FastAPI architecture requires a lot of functions calls as default arguments, so ignore it here.
-    B008,
-    # docs are missing in this project.
-    D, RST
-
-    # WPS: 3xx
-    # IMO, but the obligation to specify the base class is redundant.
-    WPS306,
-
-    # WPS: 4xx
-    # FastAPI architecture requires a lot of complex calls as default arguments, so ignore it here.
-    WPS404,
-    # again, FastAPI DI architecture involves a lot of nested functions as DI providers.
-    WPS430,
-    # used for pypika operations
-    WPS465,
-
-    # WPS: 6xx
-    # pydantic defines models in dataclasses model style, but not supported by WPS.
-    WPS601,
-extend-ignore = E203,W503
-no-accept-encodings = True
-nested-classes-whitelist=Config
-inline-quotes = double
-
-[mypy]
